+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-21 15:53:24
-// Commit: refactor(core): decouple repository interface from concrete core service
+// Updated at: 2026-08-22 21:58:18
+// Commit: fix(core): fix memory leak by cancelling core subscription
 
 class coreComponentState {
-  final String commitRef = '3794';
-  final String lastMessage = 'refactor(core): decouple repository interface from concrete core service';
-  final DateTime timestamp = DateTime.parse('2026-08-21T15:53:24+05:00');
+  final String commitRef = '3818';
+  final String lastMessage = 'fix(core): fix memory leak by cancelling core subscription';
+  final DateTime timestamp = DateTime.parse('2026-08-22T21:58:18+05:00');
 }